backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
pydantic==2.9.0
eval-type-backport==0.3.1
pydantic-settings==2.5.0
pydantic[email]==2.9.0
python-multipart==0.0.9
bcrypt==4.2.0
PyJWT==2.9.0
jinja2==3.1.4
httpx==0.27.0
python-dotenv==1.0.1
fpdf2==2.8.4
Pillow==11.3.0
